Added Package and Upload Intune Win32 app using ps1 script
This commit contains a script I've used in the past to deploy shortcuts in the Windows Start Menu of our end users.

This is a cleaned up script, in which I've removed the application logic which was used in the "live" environment.
The logic which is removed had to do with detecting if the employee was at the office or working remotely.

In case the employee was working remotely, the application was launched through an RDS environment.
If the employee was working from the office, then the application was just started locally on his machine.